feat(lib, tool): Find parameters for endpoints
We don't want to only look at the return types, we also want the
parameters to endpoints and generate those. This was missed because the
project I tested on always returned the request as part of the response
object and therefore TypeContractor added the request to what we needed
to generate. But in a differently structured application, this doesn't
necessarily happen.

This also changes `TypeChecks.ReturnsActionResult` to return `true` for
ActionResults without a generic type argument, since we now want to look
at the parameters for those methods as well.
`TypeChecks.UnwrappedReturnType` handles filtering out those that
doesn't have a return type we care about.

Additionally, every `ControllerBase` adds a ton of public methods that
returns `ActionResult` that we don't want to generate. Luckily, those
are all tagged with `[NonAction]` attributes, so filtering out those is
relatively easy to do.