feat(lib): Add smart cleanup after generating files
This keeps track of which files we've generated and finds the difference
between that list and what exists on the file system. All files and
directories that have not been touched by this run will be removed.

This should reduce the number of times `npm` freaks out when it's
running in watch mode and the tool removes the entire directory it was
watching.